Move ClassNameGlobber into instrumentor library
