+++ agent/src/main/java/com/code_intelligence/jazzer/agent/RuntimeInstrumentor.kt
@@ -15,6 +15,7 @@
 package com.code_intelligence.jazzer.agent
 
 import com.code_intelligence.jazzer.instrumentor.ClassInstrumentor
+import com.code_intelligence.jazzer.instrumentor.ClassNameGlobber
 import com.code_intelligence.jazzer.instrumentor.CoverageRecorder
 import com.code_intelligence.jazzer.instrumentor.Hook
 import com.code_intelligence.jazzer.instrumentor.InstrumentationType
@@ -23,7 +24,6 @@ import com.code_intelligence.jazzer.runtime.NativeLibHooks
 import com.code_intelligence.jazzer.runtime.TraceCmpHooks
 import com.code_intelligence.jazzer.runtime.TraceDivHooks
 import com.code_intelligence.jazzer.runtime.TraceIndirHooks
-import java.lang.IllegalArgumentException
 import java.lang.instrument.ClassFileTransformer
 import java.lang.instrument.Instrumentation
 import java.nio.file.Path
@@ -31,93 +31,6 @@ import java.security.ProtectionDomain
 import kotlin.math.roundToInt
 import kotlin.system.exitProcess
 import kotlin.time.measureTimedValue
-
-private val BASE_INCLUDED_CLASS_NAME_GLOBS = listOf(
-    "**", // everything
-)
-
-private val BASE_EXCLUDED_CLASS_NAME_GLOBS = listOf(
-    "\\[**", // array types
-    "com.code_intelligence.jazzer.**",
-    "com.sun.**", // package for Proxy objects
-    "java.**",
-    "jaz.Ter", // safe companion of the honeypot class used by sanitizers
-    "jaz.Zer", // honeypot class used by sanitizers
-    "jdk.**",
-    "kotlin.**",
-    "sun.**",
-    "javax.management.DynamicMBean", // used by ManagementFactory.getRuntimeMXBean()
-    "javax.management.NotificationEmitter", // used by ManagementFactory.getRuntimeMXBean()
-    "javax.management.NotificationBroadcaster", // used by ManagementFactory.getRuntimeMXBean()
-)
-
-class SimpleGlobMatcher(val glob: String) {
-    private enum class Type {
-        // foo.bar (matches foo.bar only)
-        FULL_MATCH,
-        // foo.** (matches foo.bar and foo.bar.baz)
-        PATH_WILDCARD_SUFFIX,
-        // foo.* (matches foo.bar, but not foo.bar.baz)
-        SEGMENT_WILDCARD_SUFFIX,
-    }
-
-    private val type: Type
-    private val prefix: String
-
-    init {
-        // Remain compatible with globs such as "\\[" that use escaping.
-        val pattern = glob.replace("\\", "")
-        when {
-            !pattern.contains('*') -> {
-                type = Type.FULL_MATCH
-                prefix = pattern
-            }
-            // Ends with "**" and contains no other '*'.
-            pattern.endsWith("**") && pattern.indexOf('*') == pattern.length - 2 -> {
-                type = Type.PATH_WILDCARD_SUFFIX
-                prefix = pattern.removeSuffix("**")
-            }
-            // Ends with "*" and contains no other '*'.
-            pattern.endsWith('*') && pattern.indexOf('*') == pattern.length - 1 -> {
-                type = Type.SEGMENT_WILDCARD_SUFFIX
-                prefix = pattern.removeSuffix("*")
-            }
-            else -> throw IllegalArgumentException(
-                "Unsupported glob pattern (only foo.bar, foo.* and foo.** are supported): $pattern"
-            )
-        }
-    }
-
-    /**
-     * Checks whether [maybeInternalClassName], which may be internal (foo/bar) or not (foo.bar), matches [glob].
-     */
-    fun matches(maybeInternalClassName: String): Boolean {
-        val className = maybeInternalClassName.replace('/', '.')
-        return when (type) {
-            Type.FULL_MATCH -> className == prefix
-            Type.PATH_WILDCARD_SUFFIX -> className.startsWith(prefix)
-            Type.SEGMENT_WILDCARD_SUFFIX -> {
-                // className starts with prefix and contains no further '.'.
-                className.startsWith(prefix) &&
-                    className.indexOf('.', startIndex = prefix.length) == -1
-            }
-        }
-    }
-}
-
-internal class ClassNameGlobber(includes: List<String>, excludes: List<String>) {
-    // If no include globs are provided, start with all classes.
-    private val includeMatchers = (if (includes.isEmpty()) BASE_INCLUDED_CLASS_NAME_GLOBS else includes)
-        .map(::SimpleGlobMatcher)
-
-    // If no include globs are provided, additionally exclude stdlib classes as well as our own classes.
-    private val excludeMatchers = (if (includes.isEmpty()) BASE_EXCLUDED_CLASS_NAME_GLOBS + excludes else excludes)
-        .map(::SimpleGlobMatcher)
-
-    fun includes(className: String): Boolean {
-        return includeMatchers.any { it.matches(className) } && excludeMatchers.none { it.matches(className) }
-    }
-}
 
 internal class RuntimeInstrumentor(
     private val instrumentation: Instrumentation,
